- reorganize folder structure - fix lint script
