+++ index.html
@@ -27,7 +27,7 @@
             <button id="start-button">Start</button>
         </div>
     </div>
-    <script src="../src/index.js" type="module"></script>
+    <script src="./index.js" type="module"></script>
 </body>
 
 </html>
