+++ package.json
@@ -5,7 +5,7 @@
   "main": "index.js",
   "scripts": {
     "dev": "http-server -c-1 -o",
-    "lint": "eslint --fix src"
+    "lint": "eslint --fix ."
   },
   "repository": {
     "type": "git",
